Save and load system for the best time survived

+++ Assets/Scripts/Main/TimesCanvas/CurrentTimeSetter.cs
@@ -5,7 +5,7 @@ using UnityEngine.UI;
 
 public class CurrentTimeSetter : MonoBehaviour
 {
-    [HideInInspector] public float currentTimeInSeconds;
+    [HideInInspector] public int currentTimeInSeconds;
 
     [SerializeField] private Text currentTimeText;
     [SerializeField] private BallSpawnerManager ballSpawnerManagerScript;
